COMMERCE-813 commerce-openapi-util - Reorganize - extract MethodImporter and extract util methods to GetterUtil
